Update attach-disk-portal.yml
I am from Microsoft Azure Customer Experience Team and recently I had received a case with the same issue where customer was unable to attach a data disk. After lot of research we could identify this was the root cause. And hence I believe it is important to have  a note for the same in our documents.